pyproject.toml
[tool.poetry]
name = "modbotdeploy"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
pymongo = "^3.11.2"
python-dotenv = "^0.15.0"
flask = "^1.1.2"
dblpy = "^0.4.0"
fastapi = "^0.63.0"
pydantic = "^1.7.3"
uvicorn = "^0.13.3"
pypresence = "^4.2.0"
youtube-dl = "^2021.5.16"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
